+++ apps/eew/sceewlog/sceewlog.py
@@ -97,6 +97,8 @@ class Listener(seiscomp.client.Application):
         self.fcmDataFile = None #File that contains keys and topic information for FCM
         
         self.eewComment = True #
+
+        self.eewScript= None 
                                
         
         
@@ -226,6 +228,23 @@ class Listener(seiscomp.client.Application):
                 self.configGetString("report.directory"))
         except:
             pass
+
+
+        try:
+            self.eewScript = self.configGetString("EEW.script")
+            seiscomp.logging.info('EEW.script = %s' % self.eewScript )
+        except Exception as e:
+            seiscomp.logging.warning('Not possible to read EEW.script. Setting this to None')
+            self.eewScript = None
+
+        if self.eewScript is not None :
+            logprefix = '/tmp/sceewlog-script.'
+            print("%s test=1 > %slog 2> %serr &" % ( self.eewScript, 
+                                                           logprefix,
+                                                           logprefix ))
+            os.system( "%s test=1 > %slog 2> %serr &" % ( self.eewScript, 
+                                                           logprefix,
+                                                           logprefix ) )
 
         return True
         
@@ -531,6 +550,8 @@ class Listener(seiscomp.client.Application):
         except Exception as e:
             seiscomp.logging.warning(
                 'ActiveMQ interface cannot be loaded: %s' % e)
+
+
         return True
 
     def generateReport(self, evID):
@@ -740,11 +761,33 @@ class Listener(seiscomp.client.Application):
             #evaluate to send or not the alert based on profiles
             self.alertEvaluation( evID, magID, updateno ) 
 
+    def execScript(self, magID, updateno):
+
+        if self.eewScript is None:
+            return        
+        
+        seiscomp.logging.debug("Sending an alert for magnitude %s " % magID)
+        orgID = self.origin_lookup[magID]
+        evID = self.event_lookup[orgID]
+
+        eew_parameters = self.event_dict[evID]['updates'][ updateno ]
+        script_args = ' '.join(["%s='%s'" % ( k, v ) for i, (k, v) in enumerate(eew_parameters.items())])
+        
+        logprefix = '/tmp/sceewlog-script.'
+        com = "%s %s > %slog 2> %serr &" % ( self.eewScript, 
+                                             script_args, 
+                                             logprefix,
+                                             logprefix ) 
+        seiscomp.logging.debug( com )
+        os.system( com )
+
     def sendAlert(self, magID):
         """
         Send an alert to a UserDisplay, if one is configured
         """
-        if self.udevt is None and self.fcm == False:
+        if (self.udevt is None and 
+            not self.fcm and 
+            not self.eewComment):
             return
         
 
@@ -816,6 +859,9 @@ class Listener(seiscomp.client.Application):
             except Exception as e:
                 seiscomp.logging.error("There was an error while notifying a comment with the Num of times that EEW alert were sent")
                 seiscomp.logging.error("Error message: %s" % repr(e))
+        
+
+
 
     def handleMagnitude(self, magnitude, parentID):
         """
@@ -849,7 +895,7 @@ class Listener(seiscomp.client.Application):
         Add picks to the cache.
         """
         try:
-            seiscomp.logging.debug("Received pick %s" % pk.publicID())
+            #seiscomp.logging.debug("Received pick %s" % pk.publicID())
             self.cache.feed(pk)
         except:
             info = traceback.format_exception(*sys.exc_info())
@@ -1264,12 +1310,15 @@ class Listener(seiscomp.client.Application):
                 self.event_dict[evID]['lastupdatesent'] = updateno 
                 
                 self.sendAlert( magID )
+                self.execScript( magID, updateno )
                 break
                             
         if len(self.profilesDic) == 0:
             #no profiles. Any origin and mag is reported
             seiscomp.logging.info('No profiles but activeMQ enabled. sending an alert...')
+            self.event_dict[evID]['lastupdatesent'] = updateno 
             self.sendAlert( magID )
+            self.execScript( magID, updateno )
                     
     def handleComment(self, comment, parentID):
         """