+++ libs/seiscomp/processing/eewamps/preprocessor.cpp
@@ -19,13 +19,13 @@
 #define SEISCOMP_COMPONENT EEWAMPS
 
 
-#include <seiscomp3/logging/log.h>
-#include <seiscomp3/processing/operator/ncomps.h>
-#include <seiscomp3/io/records/mseedrecord.h>
-#include <seiscomp3/io/recordfilter/demux.h>
-#include <seiscomp3/io/recordfilter/iirfilter.h>
-#include <seiscomp3/math/filter/iirintegrate.h>
-#include <seiscomp3/math/filter/chainfilter.h>
+#include <seiscomp/logging/log.h>
+#include <seiscomp/processing/operator/ncomps.h>
+#include <seiscomp/io/records/mseedrecord.h>
+#include <seiscomp/io/recordfilter/demux.h>
+#include <seiscomp/io/recordfilter/iirfilter.h>
+#include <seiscomp/math/filter/iirintegrate.h>
+#include <seiscomp/math/filter/chainfilter.h>
 
 #include "filter/diffcentral.h"
 #include "processors/envelope.h"
@@ -54,6 +54,8 @@ class HPreProcessor::L2Norm {
 		}
 
 		bool publish(int c) const { return c == 0; }
+
+		void reset() {}
 
 		int compIndex(const std::string &code) const {
 			if ( _parent->_streamConfig[_first].code() == code )